fix type inference issues with custom error types
the irevous solution was wrong because we only required that the input
type implement From<u32> and forced a conversion in every combinator.
But if you reused a parser that already provides the custom input type,
there was no possible conversion.

It is now made possible with a more generic conversion and removing
forced conversions. Now only the most basic parsers will force an error
type, and combinators that have their own error code will use the
unify_types trick to make sure they provide the same error code as the
underlying parsers.

It might still cause some type inference issues, but the code of
reported issues (#534, #561, #617) seems to work. We'll see during the
testing period if more of those appear.