Improve the many0 documentation
The many0 macro will fail if and only if the embedded parser
does not consume any input. It took me a while to figure this out,
so it makes sense to document it.

Also, this commit adds periods to the existing documentation, and
removes the random "0 or more" below the example.