Change Deref to return an &'anon &'a Inner instead of &'anon Inner
This change is necessary because of how Rust tracks lifetimes and
reference collapse. The Deref trait returns a reference to the interior
of a Complete struct, and does NOT copy out the interior reference. This
means that anything consuming the reference coming out of Deref::deref
loses awareness of the lifetime of the /referent/, and only knows the
lifetime of the Complete struct, which is practically always the local
scope.

Furthermore, a From impl is added to take an &'anon &'a Inner and build
a fresh Complete type with the correct lifetime parameter, NOT the 'anon
lifetime generated by Deref::deref.

As a result of this change, the following formerly-invalid code is now
valid:

```rust
let x: Complete;
let y: Complete = x.deref().into();
```

The second Complete will have the same lifetime parameter as the first,
and will not have the lifetime parameter of the first's *scope*.