use the lexical-core crate for float parsing
the previous float parsing code was first using the `recognize_float`
function, then passing the result to `str::parse()`, so input data was
parsed twice.
This commit introduces the lexical-core crate, which implements faster
float parsing and can even work in no_std. It is imported through the
"lexical" feature that is active by default