clean up the Incomplete behaviour
*breaking change*

At the beginnng, Needed was supposed to indicate the total number of
bytes needed in the input buffer, and this resulted in complicated
calculations as the parser result bubbled up through previous parsers.

This was changed to indicate only how many additional bytes were needed,
but most of th parsers did not really implement it properly, so this
commit fixes it.
Note: in some cases, like take(&str), where a char can span a varying number
of bytes, we cannot know in advance how many bytes are needed.
For take_until, we could detect that the last few bytes are a prefix of
the tag we are looking for, and indicate how many we need like for
`tag`, but I am not convinced yet that it is worth the effort (and
there's no guarantee that the prefix will be followed by the bytes we
want)