use a length check instead of PartialEq in infinite loop checks
many*, fold_many* and other related combinators have an infinite loop
check that makes sure the child parser is always consuming data
(otherwise it could be reapplied without ever stopping).
That check was done by testing if the remaining input was equal, which
consists of a length test, then checking equality of all the elements.

This commit replaces the equality check with just a length check (error
if remaining has same length as input), which should amoount to the same
work as PartialEq in the common case, but will generate less code